requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
filelock==3.15.4
Markdown==3.6
pillow==10.4.0
platformdirs==4.2.2
sqlparse==0.5.1
typing_extensions==4.12.2
virtualenv==20.26.3
